Title: 사칙연산 연습, Time: 46ms, Memory: 16MB, Status: Wrong Answer - CodeTree

+++ 250314/사칙연산 연습/practice-basic-arithmetic-operations.py
@@ -1,4 +1,4 @@
 a,b=5,30
 c=a+b
 
-print(c/8, c%8)
+print(int(c/8), int(c%8))
